renamed COT file
Got error with the name file, so changed it to "plant_modules" which worked.
